[add_acl] Assume /32 subnet if not specified
As indicated in core#76, we are not currently treating ACLs
for individual addresses and ACLs for /32 as the same entities. This can
result in invalid ACLs leftover in the metadata db.

This commit checks for a '/' in the cidr_mask, and appends '/32' if one
is not found. All of this happens after the 'any' check, so there
shouldn't be any conflicts there.